バージョン番号を更新

+++ tests/TestCase/Service/Admin/UploadFilesAdminServiceTest.php
@@ -16,6 +16,7 @@ use BcUploader\Service\Admin\UploaderFilesAdminService;
 use BcUploader\Service\Admin\UploaderFilesAdminServiceInterface;
 use BcUploader\Test\Factory\UploaderConfigFactory;
 use BcUploader\Test\Factory\UploaderFileFactory;
+use Cake\Datasource\Paging\NumericPaginator;
 use CakephpFixtureFactories\Scenario\ScenarioAwareTrait;
 
 /**
@@ -81,7 +82,7 @@ class UploadFilesAdminServiceTest extends BcTestCase
 
         //対象メソッドをコール
         $rs = $this->UploaderFilesAdminService->getViewVarsForAjaxList(
-            $this->UploaderFilesAdminService->getIndex([])->all(),
+            (new NumericPaginator())->paginate($this->UploaderFilesAdminService->getIndex([])),
             1
         );
 
@@ -99,7 +100,8 @@ class UploadFilesAdminServiceTest extends BcTestCase
     {
         //limitedフォルダーと.htaccessファイルが存在しない場合、
         $limitPath = '/var/www/html/webroot/files/uploads/limited';
-        unlink($limitPath . DS . '.htaccess');
+        if (file_exists($limitPath . DS . '.htaccess'))
+            unlink($limitPath . DS . '.htaccess');
         rmdir($limitPath);
         //対象メソッドをコール
         $rs = $this->execPrivateMethod($this->UploaderFilesAdminService, 'checkInstall', []);
